pyproject.toml
[project]
name = "research-copilot"
version = "0.1.0"
description = "意图驱动型自动案头研究工作站"
requires-python = ">=3.11"
dependencies = [
    # --- Web Framework ---
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    # --- Database ---
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # --- Agent ---
    "langgraph>=0.4.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai>=2.1.0",
    "psycopg[binary,pool]>=3.2.0",
    # --- RAG ---
    "pgvector>=0.3.6",
    "sentence-transformers>=3.4.0",
    # --- Task Queue ---
    "celery[redis]>=5.4.0",
    # --- Docker ---
    "docker>=7.1.0",
    # --- Config ---
    "pydantic-settings>=2.7.0",
    # --- Logging ---
    "structlog>=24.4.0",
    # --- HTTP ---
    "httpx>=0.28.0",
    # --- Auth ---
    "pyjwt[crypto]>=2.10.0",
    "bcrypt>=4.2.0",
    "resend>=2.0.0",
    # --- Storage ---
    "boto3>=1.35.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
    # --- Rate Limiting ---
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
optimization = [
    "dspy>=2.6.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pre-commit>=4.0.0",
    "httpx>=0.28.0",  # TestClient 依赖
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_test_loop_scope = "session"
asyncio_default_fixture_loop_scope = "session"
markers = [
    "unit: Unit tests",
    "integration: Integration tests (require external services)",
    "e2e: End-to-end tests (require full Docker Compose stack)",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
